requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.17.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.1.0
auth0-python==4.8.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.16
botocore==1.36.16
browser-use==0.1.37
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
coloredlogs==15.0.1
courlan==1.3.2
crewai==0.102.0
cryptography==44.0.1
dataclasses-json==0.6.7
dateparser==1.2.1
decorator==5.2.0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
durationpy==0.9
elevenlabs==1.11.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.17.0
filetype==1.2.0
fireworks-ai==0.15.12
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
json5==0.10.0
json_repair==0.39.0
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jusText==3.0.1
kubernetes==32.0.1
langchain==0.3.19
langchain-anthropic==0.3.3
langchain-aws==0.2.12
langchain-community==0.3.18
langchain-core==0.3.37
langchain-fireworks==0.2.7
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.6
langsmith==0.2.11
litellm==1.60.2
lmnr==0.4.59
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.20.1
openai==1.63.2
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-langchain==0.38.4
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-sqlalchemy==0.51b0
opentelemetry-instrumentation-threading==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
playwright==1.50.0
posthog==3.11.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
pyvis==0.3.2
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tokenizers==0.21.0
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.66.4
trafilatura==2.0.0
traitlets==5.14.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
uv==0.6.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
